+++ blockfrost-api/src/Blockfrost/Types/Cardano/Transactions.hs
@@ -50,6 +50,7 @@ data Transaction = Transaction
   , _transactionPoolRetireCount      :: Integer -- ^ Count of the stake pool retirement certificates within the transaction
   , _transactionAssetMintOrBurnCount :: Integer -- ^ Count of asset mints and burns within the transaction
   , _transactionRedeemerCount        :: Integer -- ^ Count of redeemers within the transaction
+  , _transactionValidContract        :: Bool    -- ^ True if this is a valid transaction, False in case of contract validation failure
   }
   deriving stock (Show, Eq, Generic)
   deriving (FromJSON, ToJSON)
@@ -78,6 +79,7 @@ instance ToSample Transaction where
       , _transactionPoolRetireCount = 0
       , _transactionAssetMintOrBurnCount = 0
       , _transactionRedeemerCount = 0
+      , _transactionValidContract = True
       }
 
 -- | Transaction input UTxO