+++ blockfrost-api/CHANGELOG.md
@@ -7,6 +7,7 @@
     * `datumHash` field changed from `Maybe Text` to `Maybe DatumHash`
 
 * Additions
+  * `Transaction` grows `Bool`ean `validContract` field
 
 # Version [0.2.1.0](https://github.com/blockfrost/blockfrost-haskell/compare/v0.2.0.0...v0.2.1.0) (2021-11-15)