+++ blockfrost-api/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * Additions
   * `Transaction` grows `Bool`ean `validContract` field
+  * Lenses for `TransactionRedeemer`
 
 # Version [0.2.1.0](https://github.com/blockfrost/blockfrost-haskell/compare/v0.2.0.0...v0.2.1.0) (2021-11-15)